+++ crates/astria-composer/tests/blackbox/main.rs
@@ -1,3 +1,4 @@
 pub mod api;
-pub mod composer;
+pub mod geth_collector;
+mod grpc_collector;
 pub mod helper;
